refactor: extract has enabled trait and enabled interface
