+++ app/Contracts/EnableInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Contracts;
+
+interface EnableInterface
+{
+    public function enable(): void;
+
+    public function disable(): void;
+}
